requirements.txt
langchain>=0.2.0
faiss-cpu
sentence-transformers
jsonlines
torch
transformers
accelerate
bitsandbytes
